// ui/src/api/model/clientRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Bank System Core API REST
 * Documentación de la API REST del Core Bancario
 * OpenAPI spec version: 0.0.1
 */
import type { SecurityQuestionAnswer } from './securityQuestionAnswer';

export interface ClientRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  nationality: string;
  birthDate: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  taxId: string;
  /**
     * @minItems 3
     * @maxItems 3
     */
  securityQuestionAnswers: SecurityQuestionAnswer[];
  /**
     * @minLength 8
     * @maxLength 2147483647
     * @pattern ^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!.]).*$
     */
  password: string;
}